Change Sender/SenderGenome range from float to uint8_t, Desc types to int, and remove Constructor::isReady

+++ source/EngineGpuKernels/Genome.cuh
@@ -269,7 +269,7 @@ struct MemoryGenome
 
 struct SenderGenome
 {
-    float range;
+    uint8_t range;
     int maxTimesSent;
 };
